requirements.txt
# Chess Neural Network Engine - Python Dependencies

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0

# Chess library
python-chess>=1.9.4

# HTTP requests for Lichess API
requests>=2.31.0

# Numerical computing
numpy>=1.24.0

# Optional but recommended
# ---------------------------------

# Progress bars
tqdm>=4.65.0

# Logging and monitoring
tensorboard>=2.13.0

# Plotting and visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data handling
pandas>=2.0.0

# Configuration management
pyyaml>=6.0

# Development tools
# ---------------------------------
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# System info
# ---------------------------------
# Python >= 3.8
# CUDA >= 11.0 (optional, for GPU training)